requirements.txt
requests 
httpx 
pydantic 
fastapi 
uvicorn 
pandas 
python-dotenv
typing-extensions
asyncio
google-generativeai
